requirements.txt
asgiref==3.4.1
boto3==1.23.10
botocore==1.26.10
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2.14
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-extensions==3.2.0
django-phonenumber-field==6.0.0
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
idna==3.3
jmespath==0.10.0
phonenumbers==8.12.52
Pillow==8.4.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.10
